Rename dir name in modeling
