+++ dygraph/ppdet/modeling/__init__.py
@@ -1,11 +1,11 @@
 from . import ops
 from . import bbox
 from . import mask
-from . import backbone
-from . import neck
-from . import head
-from . import loss
-from . import architecture
+from . import backbones
+from . import necks
+from . import heads
+from . import losses
+from . import architectures
 from . import post_process
 from . import layers
 from . import utils
@@ -13,11 +13,11 @@ from . import utils
 from .ops import *
 from .bbox import *
 from .mask import *
-from .backbone import *
-from .neck import *
-from .head import *
-from .loss import *
-from .architecture import *
+from .backbones import *
+from .necks import *
+from .heads import *
+from .losses import *
+from .architectures import *
 from .post_process import *
 from .layers import *
 from .utils import *
